+++ src/InfinitoOnboardingServiceProvider.php
@@ -104,6 +104,7 @@ class InfinitoOnboardingServiceProvider extends PackageServiceProvider
             'create_onboarding_tour_completions_table',
             'create_onboarding_tour_events_table',
             'add_translations_to_onboarding_tables',
+            'add_meta_to_onboarding_tour_completions_table',
         ];
     }
 }