+++ tests/TestCase.php
@@ -77,6 +77,7 @@ class TestCase extends Orchestra
             'create_onboarding_tour_completions_table',
             'create_onboarding_tour_events_table',
             'add_translations_to_onboarding_tables',
+            'add_meta_to_onboarding_tour_completions_table',
         ] as $migration) {
             $migration = include __DIR__ . "/../database/migrations/{$migration}.php.stub";
             $migration->up();